Bump and fix issues for NextJS bump to v15.0.2
- Resolves some overlay console errors that appear if path service
is unavailable.
- Hydration errors that were present in v14 but in v15 were throwing
errors to the console overlay in the page.
- Added an avatar local for offline to avoid overlay errors in admin
logins.
- Create unique keys for the ViewDropdown component to avoid errors
surfacing in v15.
