+++ package.json
@@ -25,7 +25,7 @@
     "@patternfly/react-table": "^6.0.0",
     "axios": "^1.7.7",
     "js-yaml": "^4.1.0",
-    "next": "^14.2.15",
+    "next": "^15.0.2",
     "next-auth": "^4.24.10",
     "node-fetch": "^3.3.2",
     "react": "18.3.1",
