+++ src/components/Contribute/Knowledge/ViewDropdown/ViewDropdown.tsx
@@ -85,10 +85,10 @@ export const ViewDropdown: React.FunctionComponent<Props> = ({ knowledgeFormData
         shouldFocusToggleOnSelect
       >
         <DropdownList>
-          <DropdownItem key="view yaml" onClick={handleViewYaml}>
+          <DropdownItem key="view-yaml" onClick={handleViewYaml}>
             <CodeIcon /> YAML Content
           </DropdownItem>
-          <DropdownItem key="view yaml" onClick={handleViewAttribution}>
+          <DropdownItem key="view-attribution" onClick={handleViewAttribution}>
             <FileIcon /> Attribution Content
           </DropdownItem>
         </DropdownList>
