+++ src/components/Contribute/Knowledge/knowledge.css
@@ -23,7 +23,7 @@
   border-color: #45a049;
 }
 
-.spinner-container{
+.spinner-container {
   display: flex;
   justify-content: center;
   align-items: center;
